main_python/requirements.txt
opencv-contrib-python==4.6.0.66
pyserial==3.5
numpy==1.26.4
sympy==1.13.2
keyboard==0.13.5
#pip install -r requirements.txt